backup/api/requirements.txt
fastapi==0.115.8
uvicorn==0.27.1
sqlmodel==0.0.14
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic[email]==2.6.1
python-dotenv==1.0.0
cryptography==42.0.5
openai==1.12.0
psycopg2-binary==2.9.9
gunicorn==21.2.0